# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recurrences:
    get:
      operationId: listRecurrences
      tags: [recurrences]
      summary: Recurring posts in the workspace
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurrenceList"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createRecurrence
      tags: [recurrences]
      summary: Create a recurring post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recurrence"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recurrences/preview:
    post:
      operationId: previewRecurrence
      tags: [recurrences]
      summary: The next dates a rule would post on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceInput"
      responses:
        "200":
          description: Dates
          content:
            application/json:
              schema:
                type: object
                properties:
                  dates:
                    type: array
                    items:
                      type: string
                      format: date-time
                required: [dates]
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recurrences/{recurrenceId}:
    parameters:
      - name: recurrenceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getRecurrence
      tags: [recurrences]
      summary: A recurring post with its occurrences
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurrenceDetail"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRecurrence
      tags: [recurrences]
      summary: Update, pause or resume a recurring post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recurrence"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecurrence
      tags: [recurrences]
      summary: Stop a recurring post and remove its unsent occurrences
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    PostLink:
      type: object
      description: A link attached to a post as a preview card; a snapshot of what the composer showed.
      properties:
        url: { type: string, maxLength: 2048 }
        title: { type: ["string", "null"], maxLength: 300 }
        description: { type: ["string", "null"], maxLength: 1000 }
        imageUrl: { type: ["string", "null"], maxLength: 2048 }
      required: [url]
...
    PostStatus:
      type: string
      enum: [draft, scheduled, publishing, published, failed]
    PostTarget:
      type: object
      properties:
        connectedAccountId:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/Platform"
        accountName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/PostStatus"
        errorMessage:
          type: ["string", "null"]
        postUrl:
          type: ["string", "null"]
          description: Link to the live post on the network, once published.
        firstCommentStatus:
          type: ["string", "null"]
          enum: [published, failed, unsupported, null]
        firstCommentError:
          type: ["string", "null"]
      required: [connectedAccountId, platform, accountName, avatarUrl, status, errorMessage, postUrl, firstCommentStatus, firstCommentError]
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        status:
          $ref: "#/components/schemas/PostStatus"
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        targets:
          type: array
          items:
            $ref: "#/components/schemas/PostTarget"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        platformContent:
          $ref: "#/components/schemas/PlatformContent"
        firstComment:
          type: ["string", "null"]
        link:
          oneOf:
            - $ref: "#/components/schemas/PostLink"
            - type: "null"
        recurrenceId:
          type: ["string", "null"]
          format: uuid
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        customValues:
          type: object
          additionalProperties: { type: string }
          description: Custom field values keyed by field ID.
      required: [id, content, status, scheduledAt, publishedAt, targets, media, platformContent, firstComment, link, recurrenceId, tags, customValues, createdAt, updatedAt]
    Media:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [image, video]
        mimeType:
          type: string
        fileName:
          type: string
        sizeBytes:
          type: integer
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        durationMs:
          type: ["integer", "null"]
        url:
          type: string
          description: Signed-in members of the workspace can GET this to view the file.
      required: [id, kind, mimeType, fileName, sizeBytes, width, height, durationMs, url]
...
    Platform:
      type: string
      enum: [facebook, instagram, linkedin, youtube]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    PlatformContent:
      type: object
      description: Per-network text overrides; a missing key means the base content is used.
      properties:
        facebook: { type: string }
        instagram: { type: string }
        linkedin: { type: string }
        youtube: { type: string }
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        postCount: { type: integer }
      required: [id, name, color]
...
    Recurrence:
      type: object
      properties:
        id: { type: string, format: uuid }
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        interval: { type: integer }
        weekdays:
          type: array
          items: { type: integer }
        dayOfMonth: { type: ["integer", "null"] }
        time: { type: string }
        timezone: { type: string }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        maxOccurrences: { type: ["integer", "null"] }
        occurrencesCreated: { type: integer }
        nextRunAt: { type: ["string", "null"], format: date-time }
        paused: { type: boolean }
        finished: { type: boolean }
        content: { type: string }
        firstComment: { type: ["string", "null"] }
        platformContent:
          $ref: "#/components/schemas/PlatformContent"
        connectedAccountIds:
          type: array
          items: { type: string, format: uuid }
        mediaIds:
          type: array
          items: { type: string, format: uuid }
        tagIds:
          type: array
          items: { type: string, format: uuid }
        upcoming:
          type: array
          items: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, frequency, interval, weekdays, dayOfMonth, time, timezone, startDate, endDate, maxOccurrences, occurrencesCreated, nextRunAt, paused, finished, content, firstComment, platformContent, connectedAccountIds, mediaIds, tagIds, upcoming, createdAt, updatedAt]
    RecurrenceDetail:
      allOf:
        - $ref: "#/components/schemas/Recurrence"
        - type: object
          properties:
            posts:
              type: array
              items:
                $ref: "#/components/schemas/Post"
          required: [posts]
    RecurrenceList:
      type: object
      properties:
        recurrences:
          type: array
          items:
            $ref: "#/components/schemas/Recurrence"
      required: [recurrences]
    RecurrenceInput:
      type: object
      properties:
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        interval: { type: integer, minimum: 1, maximum: 52 }
        weekdays:
          type: array
          items: { type: integer, minimum: 0, maximum: 6 }
        dayOfMonth: { type: ["integer", "null"], minimum: 1, maximum: 31 }
        time: { type: string }
        timezone: { type: string }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        maxOccurrences: { type: ["integer", "null"] }
        paused: { type: boolean }
        content: { type: string }
        firstComment: { type: ["string", "null"] }
        platformContent:
          $ref: "#/components/schemas/PlatformContent"
        connectedAccountIds:
          type: array
          items: { type: string, format: uuid }
        mediaIds:
          type: array
          items: { type: string, format: uuid }
        tagIds:
          type: array
          items: { type: string, format: uuid }
...